Added special check for tools those use free click.
When a user click on point the app show line between selected point and a mouse
cursor. This usefull when need to select an angle or length. But because
selection of a point also emit release event this particular event should be
ignored. In case of release inside of point circle all worked good, but not if
select point by label.

--HG--
branch : feature