Improve Find measurements location dialog.
When Valentina can't find a measurements file it asks if user want to point to
new location. In my pratcise usually a file is near the pattern file. In old way
Valentina ignored that fact and pointed to standard locations. From now one it
will first check if searched name is near the pattern file.

--HG--
branch : develop