Fix the tape app importing of measurments from a pattern
When importing measurements from a pattern in the tape app, it
defaulted to the templates directory instead of the patterns
directory. The patterns directory was set as a setting in the
valentina settings (VSettings) instead of the common settings
(VCommonSettings). I refactored the the setting into the
VCommonSettings and made the tape app use the setting. I was
able to confirm that setting a new pattern directory resulted
in the same pattern directory being used in both Valentina and
tape.

--HG--
branch : develop