Add ability to search measurements by regex.
You do this by prepending "/r/" to the front of the search string.
This makes it much easier to find measurements when one has a lot of
measurements for a given individual.

I also refactored that common table search code to consolidate this
functionality and make it easier to add other search methods in the
future.

--HG--
branch : develop