Set environment variable ICU_DATA on runtime.
When deploying with AppImage based on OpenSuse, the ICU library has a
hardcoded path to the icudt*.dat file. This prevents the library from using
shared in memory data. There are few ways to resolve this issue. According to
documentation we can either use ICU_DATA environment variable or the function
u_setDataDirectory().

--HG--
branch : develop